ml/requirements.txt
numpy>=1.24
pandas>=1.5
scikit-learn>=1.2
joblib>=1.2
torch>=2.0
sentence-transformers>=2.7
faiss-cpu>=1.8
